% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_enchist.R
\name{create_enchist}
\alias{create_enchist}
\title{Generate encounter histories}
\usage{
create_enchist(x)
}
\arguments{
\item{x}{A dataframe with columns for ResightDate/ResightPeriod and FlagID (usually generated by
\code{format_bandedbirds()} and \code{set_period} if applicable).}
}
\value{
Dataframes with encounter histories for each resighted bird that can be exported as CSV or INP files.
}
\description{
Generate encounter histories
}
\examples{
create_enchist(resight_data)
}
