% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_period.R
\name{custom_period}
\alias{custom_period}
\title{Set a custom resighting period}
\usage{
custom_period(x, y)
}
\arguments{
\item{x}{A dataframe generated by \code{format_bandedbirds} with columns for ResightDate and FlagID.}

\item{y}{A dataframe imported to R with columns for ResightDate and ResightPeriod.}
}
\value{
Dataframe with columns for ResightPeriod and FlagID (flag colour and
        code).
}
\description{
Set a custom resighting period
}
\examples{
custom_period(resight_data, resight_periods)
}
