% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_atlcan.R
\name{format_atlcan}
\alias{format_atlcan}
\title{Format Atlantic Canada Shorebirds banding data for BandedBirds.org}
\usage{
format_atlcan(x, project = y)
}
\arguments{
\item{x}{A dataframe of Atlantic Canada Shorebirds banding data}

\item{y}{The ProjectID number assigned by BandedBirds.org}
}
\value{
A dataframe of banding data formatted for submission to BandedBirds.org
}
\description{
Format Atlantic Canada Shorebirds banding data for BandedBirds.org
}
\examples{
format_atlcan(banding_data, project = 99)
}
