% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_period.R
\name{set_period}
\alias{set_period}
\title{Set a resighting period}
\usage{
set_period(x, period = NA, rsint = y)
}
\arguments{
\item{x}{A dataframe generated by \code{format_bandedbirds} with columns for ResightDate and FlagID.}

\item{period}{The duration of one resighting period (time unit used to build encounter histories).
Enter \code{"year"}, \code{"month"}, \code{"week"}, or \code{"other"}.}

\item{rsint}{If \code{period = "other"} enter the desired unit of time (i.e. \code{"3 days"}).}
}
\value{
Dataframe with columns for ResightPeriod and FlagID (flag colour and
        code).
}
\description{
Set a resighting period
}
\examples{
set_period(resight_data, period = "year")

set_period(resight_data, period = "other", rsint = "3 days")
}
