% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_bandedbirds.R
\name{format_bandedbirds}
\alias{format_bandedbirds}
\title{Tidy BandedBirds.org data}
\usage{
format_bandedbirds(x, sp = NA, cert = TRUE)
}
\arguments{
\item{x}{A dataframe formatted as submitted to BandedBirds.org.}

\item{sp}{Which species do you want to create encounter histories for? Use the four-letter AOU codes.}

\item{cert}{Was resight certainty consistently recorded? Defaults to TRUE.}
}
\value{
Dataframe with columns for ResightDate and FlagID (flag colour and
        code).
}
\description{
Tidy BandedBirds.org data
}
\examples{
format_bandedbirds(bandedbirds, sp = "REKN", cert = TRUE)
}
